deleted one include from struct_io.c
