+++ task_01/main.cpp
@@ -2,8 +2,7 @@
 #include <stdlib.h>
 #include <string.h>
 #include <time.h>
-#include "struct.h"
-#include "struct_io.h"
+#include "struct.hpp"
 
 void error_message(int code)
 {
